+++ src/index.html
@@ -9,22 +9,22 @@
   <body>
     <table id="gameBoard">
       <tr>
-        <td id="cell-0-0" onclick="handleClick(0, 0)"></td>
-        <td id="cell-0-1" onclick="handleClick(0, 1)"></td>
-        <td id="cell-0-2" onclick="handleClick(0, 2)"></td>
+        <td id="cell-0-0"></td>
+        <td id="cell-0-1"></td>
+        <td id="cell-0-2"></td>
       </tr>
       <tr>
-        <td id="cell-1-0" onclick="handleClick(1, 0)"></td>
-        <td id="cell-1-1" onclick="handleClick(1, 1)"></td>
-        <td id="cell-1-2" onclick="handleClick(1, 2)"></td>
+        <td id="cell-1-0"></td>
+        <td id="cell-1-1"></td>
+        <td id="cell-1-2"></td>
       </tr>
       <tr>
-        <td id="cell-2-0" onclick="handleClick(2, 0)"></td>
-        <td id="cell-2-1" onclick="handleClick(2, 1)"></td>
-        <td id="cell-2-2" onclick="handleClick(2, 2)"></td>
+        <td id="cell-2-0"></td>
+        <td id="cell-2-1"></td>
+        <td id="cell-2-2"></td>
       </tr>
     </table>
 
-    <script type="text/javascript" src="js/index.js"></script>
+    <script type="module" src="js/index.js" defer></script>
   </body>
 </html>
