+++ src/js/players/computer-player.js
@@ -1,0 +1,3 @@
+import Player from './player.js';
+
+export default class ComputerPlayer extends Player {}
