+++ src/js/players/human-player.js
@@ -1,0 +1,3 @@
+import Player from './player.js';
+
+export default class HumanPlayer extends Player {}
