+++ src/js/players/player.js
@@ -1,0 +1,6 @@
+export default class Player {
+  // プレイヤーの丸バツどちらか
+  constructor(symbol) {
+    this.symbol = symbol;
+  }
+}
